FIX: Event loop timers should only run for the specified time
The implementation of start_event_loop would previously just count
the number of sleeps that occurred. But this could lead to longer
event loop times if flush_events() added time into the loop. We
want the condition to be dependent on the end-time so we don't
run our loop longer than necessary.